# Documentos/Oas/Pagos/Api de pago.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Api de pago",
    "description": "Pruebas full stack tuya",
    "contact": {
      "name": "Alfonso Sarmiento Escorcia",
      "email": "manda1978@hotmail.com"
    },
    "version": "v1"
  },
  "paths": {
    "/api/Pagos/Facturar": {
      "post": {
        "tags": [
          "Pagos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacturaDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FacturaDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FacturaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Pagos/Consultar/Factura": {
      "post": {
        "tags": [
          "Pagos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultarFactura"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultarFactura"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultarFactura"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ItemFacturaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "productoId": {
            "type": "integer",
            "format": "int32"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "precio": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "FacturaDto": {
        "type": "object",
        "properties": {
          "clienteId": {
            "type": "integer",
            "format": "int32"
          },
          "productos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemFacturaDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ConsultarFactura": {
        "type": "object",
        "properties": {
          "facturaId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      }
    }
  }
}